WellMatrixMerger: introduce translation unit
